requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
supabase==2.3.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# AI and ML
openai==1.6.1
anthropic==0.7.8
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0
python-multipart==0.0.6

# Document Processing and LangExtract dependencies
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
protobuf==4.25.1
grpcio==1.60.0

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Logging and Monitoring
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Deployment
gunicorn==21.2.0
docker==6.1.3
